Using "exec" to execute the Java process from the bash itself - when launching java earlier it used to execute as a child process. Instead of creating a new process, "exec" will replace the shell process by the JVM.
Processes created directly by Docker receive the TERM signal and it is their responsibility to properly stop running child processes. That is a problem if the actual server process is spawned by a shell script, as is often the case for a Java service. In this case the TERM signal is received by the shell process, but Bash will not forward that signal to the child process. This means that the shell process will stop, but the JVM will continue to run.